+++ Changelog.md
@@ -9,6 +9,12 @@
 
 -   Update nrfutil device to v2.1.1.
 
+### Fixed
+
+-   Number input controls using `up` and `down` keys work as expected
+-   **Transmit channel** range on **Sweep Mode** no longer affect the validation
+    of **Transmit channel** on **Single** mode
+
 ## 2.2.0 - 2023-12-07
 
 ### Changed